Version 1.0 Commit
There was a bug where the canvas would reset to 0 height and 0 width when the browser was minimized. This only occurs when the canvas had a property of display: none.

Workaround is to never set the canvas display: none and simply hide the canvas when it is not selected. Successful using position fixed and visibility: hidden. Then revert to original when selected.